requirements.txt
# 深度学习框架
torch>=1.9.0
torchvision>=0.10.0
ultralytics>=8.0.0

# 图像处理
opencv-python>=4.5.0
Pillow>=8.0.0
albumentations>=1.0.0

# 数据处理
numpy>=1.19.0
pandas>=1.1.0

# 可视化
matplotlib>=3.3.0
seaborn>=0.11.0

# 工具类
PyYAML>=5.4.0
tensorboard>=2.7.0